requirements.txt
Flask==0.9
Flask-SQLAlchemy==0.16
Flask-WTF==0.8.2
Jinja2==2.6
Mako==0.7.3
MarkupSafe==0.15
SQLAlchemy==0.8.0b2
WTForms==1.0.2
Werkzeug==0.8.3
alembic==0.4.2
argparse==1.2.1
beautifulsoup4==4.1.3
icalendar==3.2
psycopg2==2.4.6
pycrypto==2.6
pytz==2012j
requests==1.1.0
wsgiref==0.1.2
